requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==11.0.3
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
python-multipart==0.0.6
requests==2.31.0
scikit-learn==1.3.0
hmmlearn==0.3.0
